Use pass instead of pattern rewriter
