+++ src/relax/ir/dataflow_pattern_functor.cc
@@ -63,10 +63,6 @@ void DFPatternVisitor::VisitDFPattern_(const DataTypePatternNode* op) {
   VisitDFPattern(op->pattern);
 }
 
-void DFPatternVisitor::VisitDFPattern_(const TargetPatternNode* op) {
-  VisitDFPattern(op->pattern);
-}
-
 void DFPatternVisitor::VisitDFPattern_(const ExprPatternNode* op) {}
 
 void DFPatternVisitor::VisitDFPattern_(const FunctionPatternNode* op) {
